pyproject.toml
[project]
name = "logforge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[uvicorn]>=0.128.7",
    "python-dotenv>=1.2.1",
    "tortoise-orm>=1.0.0",
    "tzdata>=2025.3",
    "uvicorn>=0.40.0",
]
